Python/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# MySQL Database
mysql-connector-python==8.1.0
PyMySQL==1.1.0
SQLAlchemy==2.0.21
cryptography==41.0.4

# Additional utilities
python-dotenv==1.0.0
typing-extensions==4.7.1 